added end segments to first and last slices in GoldStandardFileReader to ensure consistent solution
When we read only a subset of the gold standard, we are left with
continuations and branches from the first and to the last section that
have no corresponding end segment. Since this is not allowed according
to the continuation constraints, we now add those end segments, even
though they are not really part of the gold standard (but part of the
gold standard of the requested subset).