TED: changed handling of background label
The background label is now not treated differently from other labels. This
fixes a bug where too many cells got assigned to the background. Furthermore,
this simplifies the TED. Only for the report of errors, the background label
makes a difference and results in some splits to be false negatives, and some
merges to be false positives.

Furthermore, the sub-integral objective of the TED has been changed to the
number of pixels modified, which results in visually nicer relabeled
reconstructions.